Update service/pipeline/action_consumer.go

changed amount field name to number

updated with pr comments

added erc1155 ownership count to nfts

added sql

+++ models/events/transfer.go
@@ -14,6 +14,6 @@ type Transfer struct {
 	EventIndex        uint      `json:"event_index"`
 	SenderAddress     string    `json:"sender_address"`
 	ReceiverAddress   string    `json:"receiver_address"`
-	TokenCount        uint      `json:"count"`
+	TokenCount        uint64    `json:"count"`
 	EmittedAt         time.Time `json:"emitted_at"`
 }
