+++ service/parsers/erc1155_batch.go
@@ -52,7 +52,7 @@ func ERC1155Batch(log types.Log) ([]*events.Transfer, error) {
 			TransactionHash:   log.TxHash.Hex(),
 			SenderAddress:     log.Topics[2].Hex(),
 			ReceiverAddress:   log.Topics[3].Hex(),
-			TokenCount:        uint(count.Uint64()),
+			TokenCount:        count.Uint64(),
 			// EmmittedAt set after parsing
 		}
 		transfers = append(transfers, &transfer)
